Print available arguments that can be passed to nvBowtie()

Calling nvBowtie_usage() prints available arguments that can be passed to the nvBowtie() function.

Usage

nvBowtie_usage()

Value

An invisible Integer of call status. The value is 0 when there is not any mistakes Otherwise the value is non-zero.

References

Langmead, B., & Salzberg, S. L. (2012). Fast gapped-read alignment with Bowtie 2. Nature methods, 9(4), 357-359.

Author

Franck RICHARD

Examples

nvBowtie_usage()
#> [1] "options:"
#> [1] "General:"
#> [1] "  --verbosity         int                    [5]        verbosity level"
#> [1] "  --upto       | -u   int                    [-1]       maximum number of reads to process"
#> [1] "  --trim3      | -3   int                    [0]        trim the first N bases of 3'"
#> [1] "  --trim5      | -5   int                    [0]        trim the first N bases of 5'"
#> [1] "  --nofw                                     [false]    do not align the forward strand"
#> [1] "  --norc                                     [false]    do not align the reverse-complemented strand"
#> [1] "  --device            int                    [0]        select the given cuda device(s) (e.g. --device 0 --device 1 ...)"
#> [1] "  --file-ref                                 [false]    load reference from file"
#> [1] "  --server-ref                               [false]    load reference from server"
#> [1] "  --phred33                                  [true]     qualities are ASCII characters equal to Phred quality + 33"
#> [1] "  --phred64                                  [false]    qualities are ASCII characters equal to Phred quality + 64"
#> [1] "  --solexa-quals                             [false]    qualities are in the Solexa format"
#> [1] "  --rg-id             string                            add the RG-ID field of the SAM output header"
#> [1] "  --rg                string,val                        add an RG-TAG field of the SAM output header"
#> [1] "  --cache-writes      bool                   [false]    speed up writes on disk"
#> [1] "Paired-End:"
#> [1] "  --ff                                       [false]    paired mates are forward-forward"
#> [1] "  --fr                                       [true]     paired mates are forward-reverse"
#> [1] "  --rf                                       [false]    paired mates are reverse-forwardd"
#> [1] "  --rr                                       [false]    paired mates are reverse-reverse"
#> [1] "  --minins     |  -I  int                    [0]        minimum insert length"
#> [1] "  --maxins     |  -X  int                    [500]      maximum insert length"
#> [1] "  --overlap                                  [true]     allow overlapping mates"
#> [1] "  --no-mixed                                 [false]    only report paired alignments"
#> [1] "  --ungapped-mates | -ug                                perform ungapped mate alignment"
#> [1] "Seeding:"
#> [1] "  --seed-len   | -L   int                    [22]       seed lengths"
#> [1] "  --seed-freq  | -i   {G|L|S},x,y                       seed interval, as x + y*func(read-len) (G=log,L=linear,S=sqrt)"
#> [1] "  --max-hits          int                    [100]      maximum amount of seed hits"
#> [1] "  --max-reseed | -R   int                    [2]        number of reseeding rounds"
#> [1] "  --top               bool                   [false]    explore top seed entirely"
#> [1] "  --N                 bool                   [false]    allow substitution in seed"
#> [1] "Extension:"
#> [1] "  --mode              {best,best-exact,all}  [best]     alignment mode\n"
#> [1] "  --all        | -a                          [false]    perform all-mapping (i.e. find and report all alignments)"
#> [1] "  --local                                    [false]    perform local alignment"
#> [1] "  --rand                                     [true]     randomized seed hit selection"
#> [1] "  --no-rand                                  [false]    do not randomize seed hit selection"
#> [1] "  --max-dist          int                    [15]       maximum edit distance"
#> [1] "  --max-effort-init   int                    [15]       initial maximum number of consecutive extension failures"
#> [1] "  --max-effort | -D   int                    [15]       maximum number of consecutive extension failures"
#> [1] "  --min-ext           int                    [30]       minimum number of extensions per read"
#> [1] "  --max-ext           int                    [400]      maximum number of extensions per read"
#> [1] "  --very-fast                                           apply the very-fast presets"
#> [1] "  --fast                                                apply the fast presets"
#> [1] "  --sensitive                                           apply the sensitive presets"
#> [1] "  --very-sensitive                                      apply the very-sensitive presets"
#> [1] "  --very-fast-local                                     apply the very-fast presets"
#> [1] "  --fast-local                                          apply the fast presets"
#> [1] "  --sensitive-local                                     apply the sensitive presets"
#> [1] "  --very-sensitive-local                                apply the very-sensitive presets"
#> [1] "Scoring:"
#> [1] "  --scoring           {sw|ed|wfa}            [ed]       Smith-Waterman / Edit-Distance / Wfa scoring"
#> [1] "  --score-min         {G|L|S},x,y                       minimum score function, as x + y*func(read-len)"
#> [1] "  --ma                int                               match bonus"
#> [1] "  --mp                int,int                           mismatch min/max penalties"
#> [1] "  --np                int                               N penalty"
#> [1] "  --rdg               int,int                           read open/extension gap penalties"
#> [1] "  --rfg               int,int                           reference open/extension gap penalties"
#> [1] "Alternative:"
#> [1] "  --wfa                                                 Activate wavefront algorithm"
#> [1] "Reporting:"
#> [1] "  --mapQ-filter | -Q  int                    [0]        minimum mapQ threshold"
#> [1] ""
#> [1] ""
#> [1] "Default values are indicated in brackets []."
#> [1] ""
#> [1] "The '--scoring-scheme filename' option allows to provide a custom Smith-Waterman scoring"
#> [1] "scheme through a text file, where each line must contain a token value pair."
#> [1] " The tokens and default values are reported below:"
#> [1] "*  match               0        // local alignment: 2"
#> [1] "*  mm-penalty-min      2"
#> [1] "*  mm-penalty-max      6"
#> [1] "*  N-penalty-min       1"
#> [1] "*  N-penalty-max       1"
#> [1] "*  score-min-const     -0.6     // local alignment: 0"
#> [1] "*  score-min-coeff     -0.6     // local alignment: 10"
#> [1] "*  score-min-type      linear   // local alignment: log"
#> [1] "*  N-ceil-const        0"
#> [1] "*  N-ceil-coeff        0.15"
#> [1] "*  read-gap-const      5"
#> [1] "*  read-gap-coeff      3"
#> [1] "*  ref-gap-const       5"
#> [1] "*  ref-gap-coeff       3"
#> [1] "*  gap-free            5"